+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 
 - Replace SwiftShell with Shell https://github.com/tuist/simulator/pull/10 by @pepibumur
 
+### Added
+
+- `Device.reload` method https://github.com/tuist/simulator/pull/11 by @pepibumur.
+- `Device.wait` method https://github.com/tuist/simulator/pull/11 by @pepibumur.
+
 ## 0.3.0
 
 ### Added